perf: Consolidate duplicative editor warmups
Warming and preloading the editor in both the MySiteViewController and
BlogDashboardViewModel is redundant. The former embeds the latter.
